Refactored to clarify non-bean usage of JwtDecoder.
